WE now loads video from youtube search
Wallpaper Engine now load the video from youtube search. How this is
acheived is by using phantomJS to parse the youtube search page of the
song title and artist. Then this is passed by the websocket. The videos
are set to auto play however they are not muted and are not synced to
the song that is already playing. I will address these with the chagne
to the proper youtube iframe API.